# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Радар вузов API
  description: '

    CRM контроля взаимодействия ИТ Школы Ростелекома с вузами.


    Все методы `/api/v1` требуют `Authorization: Bearer <access_token>` из Keycloak.

    Ошибки возвращаются в формате `application/problem+json` с полем `code`

    из каталога в ARCHITECTURE.md.

    '
  version: 0.1.0
paths:
...
  /api/v1/analytics/rating:
    get:
      tags:
      - analytics
      summary: Рейтинг востребованности
      description: Балл считается по заявкам, обучающимся и потокам с нормированием внутри направления.
        Вклад каждой метрики возвращается вместе с баллом, а неполные данные помечаются.
      operationId: read_rating_api_v1_analytics_rating_get
      security:
      - HTTPBearer: []
      parameters:
      - name: entity
        in: query
        required: false
        schema:
          enum:
          - program
          - university
          type: string
          description: Что сравниваем
          default: program
          title: Entity
        description: Что сравниваем
      - name: period_from
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Начало периода
          title: Period From
        description: Начало периода
      - name: period_to
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          description: Конец периода
          title: Period To
        description: Конец периода
      - name: direction_id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
              format: uuid
          - type: 'null'
          description: ИТ-направление
          title: Direction Id
        description: ИТ-направление
      - name: order
        in: query
        required: false
        schema:
          enum:
          - score
          - priority
          type: string
          description: 'Порядок строк: по баллу или по ручному приоритету'
          default: score
          title: Order
        description: 'Порядок строк: по баллу или по ручному приоритету'
      - name: w_applications
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 100
            minimum: 0
          - type: 'null'
          title: W Applications
      - name: w_students
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 100
            minimum: 0
          - type: 'null'
          title: W Students
      - name: w_streams
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 100
            minimum: 0
          - type: 'null'
          title: W Streams
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatingOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '304':
          description: С прошлого запроса ничего не изменилось
  /api/v1/analytics/rating/weights:
    get:
      tags:
      - analytics
      summary: Веса рейтинга по умолчанию
      operationId: read_weights_api_v1_analytics_rating_weights_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeightsOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
    put:
      tags:
      - analytics
      summary: Изменить веса рейтинга
      description: Сумма весов — 100. Меняют руководитель и администратор.
      operationId: put_weights_api_v1_analytics_rating_weights_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WeightsUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeightsOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: AUTH_FORBIDDEN
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: VALIDATION_ERROR
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - HTTPBearer: []
  /api/v1/analytics/stats/funnel:
    get:
      tags:
      - analytics
      summary: Воронка по этапам процесса группы
      operationId: read_funnel_api_v1_analytics_stats_funnel_get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Группа контрагентов; по умолчанию — вузы
          title: Group Id
        description: Группа контрагентов; по умолчанию — вузы
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/analytics/stats/stage-durations:
    get:
      tags:
      - analytics
      summary: Средняя длительность этапов процесса группы
      operationId: read_stage_durations_api_v1_analytics_stats_stage_durations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Группа контрагентов; по умолчанию — вузы
          title: Group Id
        description: Группа контрагентов; по умолчанию — вузы
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: NOT_FOUND
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/analytics/stats/distribution:
    get:
      tags:
      - analytics
      summary: Распределение по направлениям
      description: Без группы — по всем группам сразу.
      operationId: read_distribution_api_v1_analytics_stats_distribution_get
      security:
      - HTTPBearer: []
      parameters:
      - name: group_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Группа контрагентов
          title: Group Id
        description: Группа контрагентов
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartOut'
        '401':
          description: AUTH_REQUIRED
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: Validation Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
...
    ChartOut:
      properties:
        title:
          type: string
          title: Title
        labels:
          items:
            type: string
          type: array
          title: Labels
        values:
          items:
            type: integer
          type: array
          title: Values
        option:
          additionalProperties: true
          type: object
          title: Option
          description: Готовые настройки ECharts
      type: object
      required:
      - title
      - labels
      - values
      - option
      title: ChartOut
      description: 'Данные и настройки ECharts: интерфейс и PDF рисуют график по одному описанию.'
...
    ContributionOut:
      properties:
        metric:
          type: string
          title: Metric
          description: applications, students или streams
        weight:
          type: integer
          title: Weight
        value:
          type: number
          title: Value
          description: Значение метрики за период
        normalized:
          type: number
          title: Normalized
          description: Доля от лучшего в направлении, 0–100
        contribution:
          type: number
          title: Contribution
          description: Вклад метрики в балл
      type: object
      required:
      - metric
      - weight
      - value
      - normalized
      - contribution
      title: ContributionOut
...
    ErrorCode:
      type: string
      enum:
      - VALIDATION_ERROR
      - AUTH_REQUIRED
      - AUTH_FORBIDDEN
      - NOT_FOUND
      - INTERACTION_VERSION_CONFLICT
      - INTERACTION_DUPLICATE
      - WF_TRANSITION_NOT_ALLOWED
      - WF_COMMENT_REQUIRED
      - WF_ATTACHMENT_REQUIRED
      - WF_VERSION_NOT_DRAFT
      - WF_MIGRATION_MAP_INCOMPLETE
      - FILE_TYPE_NOT_ALLOWED
      - FILE_TOO_LARGE
      - IMPORT_MAPPING_INVALID
      - REPORT_TOO_LARGE
      - INTEGRATION_UNAVAILABLE
      - INTERNAL_ERROR
      title: ErrorCode
...
    FieldError:
      properties:
        field:
          type: string
          title: Field
        message:
          type: string
          title: Message
      type: object
      required:
      - field
      - message
      title: FieldError
...
    Problem:
      properties:
        type:
          type: string
          title: Type
          default: about:blank
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          $ref: '#/components/schemas/ErrorCode'
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
        trace_id:
          type: string
          title: Trace Id
        errors:
          items:
            $ref: '#/components/schemas/FieldError'
          type: array
          title: Errors
          default: []
      type: object
      required:
      - title
      - status
      - code
      - trace_id
      title: Problem
      description: Тело ответа с ошибкой.
...
    RatingOut:
      properties:
        entity:
          type: string
          enum:
          - program
          - university
          title: Entity
        weights:
          additionalProperties:
            type: integer
          type: object
          title: Weights
        order:
          type: string
          enum:
          - score
          - priority
          title: Order
          description: 'Чем отсортированы строки: баллом или ручным приоритетом'
          default: score
        rows:
          items:
            $ref: '#/components/schemas/RatingRowOut'
          type: array
          title: Rows
      type: object
      required:
      - entity
      - weights
      - rows
      title: RatingOut
    RatingRowOut:
      properties:
        place:
          type: integer
          title: Place
        place_change:
          anyOf:
          - type: integer
          - type: 'null'
          title: Place Change
          description: Насколько поднялась позиция к предыдущему периоду; null — раньше её не было
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        direction_name:
          type: string
          title: Direction Name
        score:
          type: number
          title: Score
        priority:
          type: integer
          title: Priority
          description: Ручной приоритет курса, 0 — не задан
        contributions:
          items:
            $ref: '#/components/schemas/ContributionOut'
          type: array
          title: Contributions
        complete:
          type: boolean
          title: Complete
          description: Есть ли все три метрики за период
        missing_metrics:
          items:
            type: string
          type: array
          title: Missing Metrics
      type: object
      required:
      - place
      - place_change
      - id
      - name
      - direction_name
      - score
      - priority
      - contributions
      - complete
      - missing_metrics
      title: RatingRowOut
...
    WeightsOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        w_applications:
          type: integer
          title: W Applications
        w_students:
          type: integer
          title: W Students
        w_streams:
          type: integer
          title: W Streams
        is_default:
          type: boolean
          title: Is Default
      type: object
      required:
      - id
      - name
      - w_applications
      - w_students
      - w_streams
      - is_default
      title: WeightsOut
    WeightsUpdate:
      properties:
        w_applications:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: W Applications
        w_students:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: W Students
        w_streams:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: W Streams
      type: object
      required:
      - w_applications
      - w_students
      - w_streams
      title: WeightsUpdate
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: Access token из Keycloak
      scheme: bearer